requirements.txt
# ══════════════════════════════════════════════════════════════════
# Taskware Manager - Dependencies (Linux Only)
# Malware-Centric Process Monitor
# ══════════════════════════════════════════════════════════════════

# Core GUI Framework
PyQt6>=6.5.0,<7.0.0

# System Monitoring
psutil>=5.9.0,<6.0.0

# YARA Scanning
yara-python>=4.3.0,<5.0.0

# File System Monitoring
watchdog>=3.0.0,<5.0.0

# ML Model Dependencies
joblib>=1.3.0,<2.0.0
pandas>=2.0.0,<3.0.0
numpy>=1.24.0,<2.0.0
scikit-learn>=1.3.0,<2.0.0

# File Type Detection (optional, falls back to `file` command)
python-magic>=0.4.27,<1.0.0

# HTTP (for Hybrid Analysis API, optional)
requests>=2.31.0,<3.0.0

# Already in stdlib: sqlite3, hashlib, math, socket, subprocess, re, json
